Refactor code style and add paging

+++ include/pmm.h
@@ -4,10 +4,10 @@
 #include <stdint.h>
 #include <stddef.h>
 
-#define PMM_BLOCK_SIZE        4096     // 4KB blocks
-#define PMM_BLOCKS_PER_BYTE   8        // 8 blocks per byte (1 bit per block)
+#define PMM_BLOCK_SIZE 4096   // 4KB blocks
+#define PMM_BLOCKS_PER_BYTE 8 // 8 blocks per byte (1 bit per block)
 
-void pmm_init(size_t mem_size, uint32_t *bitmap);
+void pmm_init(size_t mem_size, uint8_t *bitmap); // changed type from uint32_t* to uint8_t*
 void pmm_mark_used_region(uint32_t base, uint32_t size);
 void *pmm_alloc_block();
 void *pmm_alloc_blocks(int num_blocks);
